refactor(data-access): move guide-template/asset I/O onto the canonical boundary
The two client-side data modules under src/data (guide-templates/supabase-client
and guide-assets/supabase-client) held all 21 grandfathered no-restricted-syntax
warnings — they called supabase .from()/.storage directly, which the data-access
boundary rule reserves for src/lib/supabase/<domain>. Relocate them verbatim to
src/lib/supabase/guide-templates.ts and src/lib/supabase/guide-assets.ts (a pure
move — src/lib/supabase is the allowed location and neither imports @/features) and
repoint every importer. Logic unchanged.

Boundary detector now reports zero violations: lint is 0 errors / 0 warnings
(was 0/21); the ratchet baseline is reset to 0 so drift can only be caught, never
grandfathered. Full suite 1359 green.
