+++ .gitignore
@@ -95,3 +95,6 @@ ds-bundle/
 .design-sync/learnings/
 .design-sync/node_modules
 .design-sync/previews/.cache/
+
+# Generated Graphify graph cache (local analysis artifact)
+graphify-out/
