+++ src/lib/supabase/guide-assets.ts
