+++ src/lib/supabase/guide-templates.ts
